Removed warnings over signedness.

+++ beancount/ccore/number.cc
@@ -71,7 +71,7 @@ std::ostream& operator<<(std::ostream& os, const Number& proto) {
 
 void CopySansCommas(const char* src, char* buffer, size_t num_chars) {
   char* dst = buffer;
-  for (int i = 0; i < num_chars; ++i, ++src) {
+  for (size_t i = 0; i < num_chars; ++i, ++src) {
     if (*src == ',')
       continue;
     *dst++ = *src;
